+++ src/websockets/server.py
@@ -667,7 +667,7 @@ class Serve:
     :class:`WebSocketServerProtocol` instance. It defaults to
     :class:`WebSocketServerProtocol`.
 
-    The behavior of the ``ping_interval``, ``ping_timeout``, ``timeout``,
+    The behavior of the ``ping_interval``, ``ping_timeout``, ``close_timeout``,
     ``max_size``, ``max_queue``, ``read_limit``, and ``write_limit`` optional
     arguments is described in the documentation of
     :class:`~websockets.protocol.WebSocketCommonProtocol`.
@@ -721,14 +721,15 @@ class Serve:
         create_protocol=None,
         ping_interval=20,
         ping_timeout=20,
-        timeout=10,
+        close_timeout=None,
         max_size=2 ** 20,
         max_queue=2 ** 5,
         read_limit=2 ** 16,
         write_limit=2 ** 16,
         loop=None,
         legacy_recv=False,
-        klass=None,
+        klass=WebSocketServerProtocol,
+        timeout=10,
         origins=None,
         extensions=None,
         subprotocols=None,
@@ -736,13 +737,15 @@ class Serve:
         compression='deflate',
         **kwds
     ):
+        # Backwards-compatibility: close_timeout used to be called timeout.
+        # If both are specified, timeout is ignored.
+        if close_timeout is None:
+            close_timeout = timeout
+
         # Backwards-compatibility: create_protocol used to be called klass.
-        # In the unlikely event that both are specified, klass is ignored.
+        # If both are specified, klass is ignored.
         if create_protocol is None:
             create_protocol = klass
-
-        if create_protocol is None:
-            create_protocol = WebSocketServerProtocol
 
         if loop is None:
             loop = asyncio.get_event_loop()
@@ -770,7 +773,7 @@ class Serve:
             secure=secure,
             ping_interval=ping_interval,
             ping_timeout=ping_timeout,
-            timeout=timeout,
+            close_timeout=close_timeout,
             max_size=max_size,
             max_queue=max_queue,
             read_limit=read_limit,